Extract resharding driver stages into separate modules
